+++ Inteligencia_Computacional/Autimatic/index.html
@@ -267,5 +267,5 @@
         <p>Todos os direitos reservados - JVVN</p>
     </footer>
 </body>
-<script src="./js/script.js"></script>
+<script type="module" src="./js/script.js"></script>
 </html>
